add cloudinary

+++ backend/controllers/foodController.js
@@ -1,5 +1,6 @@
 import foodModel from "../models/foodModel.js";
-import fs from "fs";
+import uploadsToCloudinary from "../helper/cloudinayHelper.js"; 
+import cloudinary from "../config/cloudinary.js";
 
 // all food list
 const listFood = async (req, res) => {
@@ -15,14 +16,14 @@ const listFood = async (req, res) => {
 // add food
 const addFood = async (req, res) => {
   try {
-    let image_filename = `${req.file.filename}`;
+    const image = await uploadsToCloudinary(req.file.buffer);
 
     const food = new foodModel({
       name: req.body.name,
       description: req.body.description,
       price: req.body.price,
       category: req.body.category,
-      image: image_filename,
+      image: [image.secure_url, image.public_id],
     });
 
     await food.save();
@@ -37,7 +38,10 @@ const addFood = async (req, res) => {
 const removeFood = async (req, res) => {
   try {
     const food = await foodModel.findById(req.body.id);
-    fs.unlink(`uploads/${food.image}`, () => {});
+    if(!food) {
+      return res.json({ success: false, message: "Food Not Found" });
+    }
+    await cloudinary.uploader.destroy(food.image[1]);
 
     await foodModel.findByIdAndDelete(req.body.id);
     res.json({ success: true, message: "Food Removed" });
